+++ app/websocket/src/main/kotlin/com/xorker/draw/websocket/broker/WebSocketBroadcaster.kt
@@ -1,11 +1,11 @@
-package com.xorker.draw.websocket.event
+package com.xorker.draw.websocket.broker
 
 import com.xorker.draw.websocket.BroadcastEvent
 import org.springframework.context.ApplicationEventPublisher
 import org.springframework.stereotype.Component
 
 @Component
-class EventPublisher(
+class WebSocketBroadcaster(
     private val publisher: ApplicationEventPublisher,
 ) {
     fun publishBroadcastEvent(event: BroadcastEvent) {
