+++ app/websocket/src/main/kotlin/com/xorker/draw/websocket/message/response/MafiaGameMessengerImpl.kt
@@ -4,7 +4,7 @@ import com.xorker.draw.mafia.MafiaGameMessenger
 import com.xorker.draw.mafia.MafiaPlayer
 import com.xorker.draw.room.Room
 import com.xorker.draw.websocket.BroadcastEvent
-import com.xorker.draw.websocket.event.EventPublisher
+import com.xorker.draw.websocket.broker.WebSocketBroadcaster
 import com.xorker.draw.websocket.message.response.dto.MafiaPlayerListBody
 import com.xorker.draw.websocket.message.response.dto.MafiaPlayerListMessage
 import com.xorker.draw.websocket.message.response.dto.toResponse
@@ -12,7 +12,7 @@ import org.springframework.stereotype.Component
 
 @Component
 class MafiaGameMessengerImpl(
-    private val eventPublisher: EventPublisher,
+    private val broadcaster: WebSocketBroadcaster,
 ) : MafiaGameMessenger {
     override fun broadcastPlayerList(room: Room<MafiaPlayer>) {
         val roomId = room.id
@@ -26,6 +26,6 @@ class MafiaGameMessengerImpl(
 
         val event = BroadcastEvent(roomId, message)
 
-        eventPublisher.publishBroadcastEvent(event)
+        broadcaster.publishBroadcastEvent(event)
     }
 }
